update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@53771 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-4.4/fs/cifs/smb2ops.c
@@ -616,7 +616,7 @@ smb2_clone_range(const unsigned int xid,
 		pcchunk->SourceOffset = cpu_to_le64(src_off);
 		pcchunk->TargetOffset = cpu_to_le64(dest_off);
 		pcchunk->Length =
-			cpu_to_le32(min_t(u32, len, tcon->max_bytes_chunk));
+			cpu_to_le32(min_t(u64, len, tcon->max_bytes_chunk));
 
 		/* Request server copy to target from src identified by key */
 		kfree(retbuf);
